added css_selector for incorrect colid error message and merged tests into one file
